Land http://codereview.chromium.org/509029 (slightly simplified).
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@3522 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/x64/fast-codegen-x64.cc
@@ -1678,8 +1678,7 @@ Register FastCodeGenerator::context_register() { return rsi; }
 
 
 void FastCodeGenerator::StoreToFrameField(int frame_offset, Register value) {
-  ASSERT_EQ(POINTER_SIZE_ALIGN(frame_offset),
-            static_cast<intptr_t>(frame_offset));
+  ASSERT_EQ(POINTER_SIZE_ALIGN(frame_offset), frame_offset);
   __ movq(Operand(rbp, frame_offset), value);
 }
